pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cicd-fixer-agent"
version = "0.1.0"
description = "AI-powered CI/CD failure analysis and fix generation system"
authors = [
    {name = "CI/CD Fixer Team", email = "team@cicdfixer.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.35.0",
    "requests>=2.32.5",
    "python-multipart>=0.0.20",
    "google-genai>=1.31.0",
    "python-dotenv>=1.1.1",
    "pydantic>=2.11.7",
    "sqlalchemy>=2.0.43",
    "psycopg2-binary>=2.9.10",
    "portia-sdk-python[google]>=0.7.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]

[project.urls]
Homepage = "https://github.com/your-org/cicd-fixer-agent"
Repository = "https://github.com/your-org/cicd-fixer-agent"
Documentation = "https://cicd-fixer-agent.readthedocs.io"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
